ETS-Server/mqtt/requirements.txt
pandas
numpy
mysql-connector
pyyaml
paho-mqtt